splitting PipeSpec into multiple files, moving to streaming
